Slight changes to the character controller logic
Not everything is completely put together yet, and I'm not sure if this is
what you had in mind for the KeyEvents (keyPressed vs. keyReleased), but I
have it set up now specifically for the character running that if the
player holds down the R button, the character runs and the character speed
should be updated accordingly, and the character's speed is set back to the
default speed once the R button is released. I'm assuming this is what
we want.