Character doesn't run unless already moving
The 'Run' key listener shouldn't interact with the character controller
unless the player is actually moving now. Tried a slightly different
approach with how the other key events interact with the controller
(though everything should be functionally exactly the same as before).
Still need to fix the character's updated position when movement keys
are pressed.